simplify endpoints
