+++ src/Halcyon.Api/Features/Tags.cs
@@ -3,8 +3,7 @@
 public static class Tags
 {
     public const string Account = "Account";
-    public const string Manage = "Manage";
+    public const string Profile = "Profile";
     public const string Seed = "Seed";
-    public const string Token = "Token";
     public const string Users = "Users";
 }
